requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
apispec==6.6.1
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
click==8.1.7
distro==1.9.0
dominate==2.9.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
flask-smorest==0.44.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
msgspec==0.18.6
openai==1.35.10
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.12.2
visitor==0.1.3
webargs==8.4.0
Werkzeug==3.0.3
WTForms==3.1.2
flask-limiter==3.7.0
pydub==0.25.1   # For audio processing
openai-whisper  # OpenAI Whisper STT library
soundfile       # Required by whisper
requests
elevenlabs==1.7.0
boto3
